refactor(sosmed): gabung 4 helper jadi satu sosmed_helper
